move test to the independent test file to align the fixtures
