+++ client/e2e/05 - custom-codes.spec.ts
@@ -1,13 +1,4 @@
 import { test, expect } from './fixtures/fixtures';
-import {
-  createAndActivateCovidConfig,
-  deleteConfigurationArtifacts,
-} from './utils';
-import path from 'path';
-
-const ESSENTIAL_HYPERTENSION_SNOMED = '59621000';
-const EXPECTED_HYPERTENSION_ELEMENT =
-  '<value code="59621000" codeSystem="2.16.840.1.113883.6.96" codeSystemName="SNOMED-CT" displayName="Essential hypertension (disorder)" xsi:type="CD"/>';
 
 const makeCsvContent = (
   rows: { code: string; system: string; name: string }[]
@@ -179,136 +170,5 @@ test.describe('Custom code builder flows', () => {
       page.getByRole('button', { name: 'Import from CSV' })
     ).toBeVisible();
     await expect(makeAxeBuilder).toHaveNoAxeViolations();
-  });
-
-  test('adds a custom code and checks the value gets refined', async ({
-    page,
-    makeAxeBuilder,
-  }) => {
-    await expect(makeAxeBuilder).toHaveNoAxeViolations();
-
-    // use the COVID configuration, or make it if it doesn't exist
-    await page.goto('/configurations');
-
-    await expect(
-      page.getByRole('heading', { name: 'Configurations' })
-    ).toBeVisible();
-
-    let covidConfig = page.locator('[data-testid="COVID-19-button"]');
-    const covidButtonCount = await covidConfig.count();
-
-    let createdNewConfig = false;
-
-    if (covidButtonCount === 0) {
-      await createAndActivateCovidConfig(page);
-      createdNewConfig = true;
-    }
-
-    await page.goto('/configurations');
-    covidConfig = page.getByTestId('COVID-19-button');
-
-    await covidConfig.click();
-    await expect(
-      page.getByRole('heading', { name: 'COVID-19', exact: true })
-    ).toBeVisible();
-
-    let someDraftExists = false;
-    const goToDraftButton = page.locator('[data-testid="go-to-draft-button"]');
-    const goToDraftButtonCount = await goToDraftButton.count();
-
-    if (goToDraftButtonCount > 0) {
-      someDraftExists = true;
-      await goToDraftButton.click();
-    }
-
-    const createDraftButton = page.locator(
-      '[data-testid="create-draft-button"]'
-    );
-    const createToDraftButtonCount = await createDraftButton.count();
-
-    if (createToDraftButtonCount > 0) {
-      someDraftExists = true;
-      await createDraftButton.click();
-      await page.getByText('Yes, draft a new version').click();
-    }
-
-    const hyperTensionCode = page.locator(
-      `data-testid=${ESSENTIAL_HYPERTENSION_SNOMED}`
-    );
-    await expect(
-      page.getByRole('heading', { name: 'Build configuration' })
-    ).toBeVisible();
-
-    await page.getByRole('button', { name: 'Custom codes' }).click();
-    await expect(
-      page.getByRole('heading', { name: 'Custom codes' })
-    ).toBeVisible();
-
-    const hyperTensionCodeCount = await hyperTensionCode.count();
-    if (hyperTensionCodeCount === 0) {
-      await createHypertensionCode();
-    }
-
-    await page.getByText('Activate').click();
-
-    if (someDraftExists) {
-      await page.getByRole('button', { name: 'Switch to version' }).click();
-      await page
-        .getByRole('button', { name: 'Yes, switch to Version' })
-        .click();
-    }
-
-    await expect(page.getByText('Configuration activated')).toBeVisible();
-
-    await page.getByRole('link', { name: 'Testing' }).click();
-    await expect(page.getByText('Test Refiner')).toBeVisible();
-
-    const fileInput = page.locator('input#zip-upload');
-
-    // Resolve the file path relative to the project root
-    const filePath = path.resolve(
-      process.cwd(),
-      'e2e/assets/mon-mothma-two-conditions.zip'
-    );
-
-    await fileInput.setInputFiles(filePath);
-
-    await page.getByText('Refine .zip file').click();
-    await page.getByText('Refine eCR').click();
-
-    await expect(
-      page.getByRole('heading', { name: 'eCR refinement results' })
-    ).toBeVisible();
-
-    await page.getByRole('button', { name: 'Show stacked diff' }).click();
-
-    // hypertension element should be retained and show up exactly once
-    await expect(page.getByText(EXPECTED_HYPERTENSION_ELEMENT)).toBeVisible();
-
-    // tear down made artifacts if we created it
-    if (createdNewConfig) {
-      deleteConfigurationArtifacts('COVID-19');
-    }
-
-    async function createHypertensionCode() {
-      await page.getByRole('button', { name: 'Custom codes' }).click();
-      await expect(
-        page.getByRole('heading', { name: 'Custom codes', level: 3 })
-      ).toBeVisible();
-      await page.getByRole('button', { name: 'Add new custom code' }).click();
-      const submitButton = page.getByRole('button', {
-        name: 'Add custom code',
-      });
-      await page
-        .getByRole('textbox', { name: 'Code #' })
-        .fill(ESSENTIAL_HYPERTENSION_SNOMED);
-      await page.getByLabel('Code system').selectOption('snomed');
-      await page
-        .getByRole('textbox', { name: 'Code name' })
-        .fill('Essential Hypertension');
-
-      await submitButton.click();
-      await expect(page.getByText('Custom code added')).toBeVisible();
-    }
   });
 });
